services/background-removal/requirements.txt
rembg[gpu]>=2.0.50
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0